@@ -1,67 +1,64 @@
-namespace DesafioFundamentos.Models
+namespace DesafioFundamentos.Models;
+public class Estacionamento
 {
-    public class Estacionamento
+    private decimal precoInicial = 0;
+    private decimal precoPorHora = 0;
+    private List<string> veiculos = new List<string>();
+
+    public Estacionamento(decimal precoInicial, decimal precoPorHora)
     {
-        private decimal precoInicial = 0;
-        private decimal precoPorHora = 0;
-        private List<string> veiculos = new List<string>();
+        this.precoInicial = precoInicial;
+        this.precoPorHora = precoPorHora;
+    }
 
-        public Estacionamento(decimal precoInicial, decimal precoPorHora)
+    public void AdicionarVeiculo()
+    {
+        Console.WriteLine("Digite a placa do veículo para estacionar:");
+        veiculos.Add(Console.ReadLine());
+    }
+
+    public void RemoverVeiculo()
+    {
+        Console.WriteLine("Digite a placa do veículo para remover:");
+
+        string placa = "";
+        placa = Console.ReadLine();
+
+        if (!veiculos.Exists(x => x.ToUpper() == placa.ToUpper()))
         {
-            this.precoInicial = precoInicial;
-            this.precoPorHora = precoPorHora;
+            Console.WriteLine("Desculpe, esse veículo não está estacionado aqui. Confira se digitou a placa corretamente");
+            return;
+        }
+        
+        Console.WriteLine("Digite a quantidade de horas que o veículo permaneceu estacionado:");
+
+        var horasInput = Console.ReadLine();
+        var convertidoComSucesso = int.TryParse(horasInput, out int horas);
+        
+        if (!convertidoComSucesso || horas < 0)
+        {
+            Console.WriteLine("Valor inválido, tente novamente");
+            return;
         }
 
-        public void AdicionarVeiculo()
+        decimal valorTotal = precoInicial + (precoPorHora * horas);
+
+        veiculos.Remove(placa);
+        Console.WriteLine($"O veículo {placa} foi removido e o preço total foi de: R$ {valorTotal}");
+    }
+
+    public void ListarVeiculos()
+    {
+        if (!veiculos.Any())
         {
-            // TODO: Pedir para o usuário digitar uma placa (ReadLine) e adicionar na lista "veiculos"
-            // *IMPLEMENTE AQUI*
-            Console.WriteLine("Digite a placa do veículo para estacionar:");
+            Console.WriteLine("Não há veículos estacionados.");
+            return;
         }
 
-        public void RemoverVeiculo()
+        Console.WriteLine("Os veículos estacionados são:");
+        for (int i = 0; i < veiculos.Count; i++)
         {
-            Console.WriteLine("Digite a placa do veículo para remover:");
-
-            // Pedir para o usuário digitar a placa e armazenar na variável placa
-            // *IMPLEMENTE AQUI*
-            string placa = "";
-
-            // Verifica se o veículo existe
-            if (veiculos.Any(x => x.ToUpper() == placa.ToUpper()))
-            {
-                Console.WriteLine("Digite a quantidade de horas que o veículo permaneceu estacionado:");
-
-                // TODO: Pedir para o usuário digitar a quantidade de horas que o veículo permaneceu estacionado,
-                // TODO: Realizar o seguinte cálculo: "precoInicial + precoPorHora * horas" para a variável valorTotal                
-                // *IMPLEMENTE AQUI*
-                int horas = 0;
-                decimal valorTotal = 0; 
-
-                // TODO: Remover a placa digitada da lista de veículos
-                // *IMPLEMENTE AQUI*
-
-                Console.WriteLine($"O veículo {placa} foi removido e o preço total foi de: R$ {valorTotal}");
-            }
-            else
-            {
-                Console.WriteLine("Desculpe, esse veículo não está estacionado aqui. Confira se digitou a placa corretamente");
-            }
-        }
-
-        public void ListarVeiculos()
-        {
-            // Verifica se há veículos no estacionamento
-            if (veiculos.Any())
-            {
-                Console.WriteLine("Os veículos estacionados são:");
-                // TODO: Realizar um laço de repetição, exibindo os veículos estacionados
-                // *IMPLEMENTE AQUI*
-            }
-            else
-            {
-                Console.WriteLine("Não há veículos estacionados.");
-            }
+            Console.WriteLine(veiculos[i]);
         }
     }
 }
